added Indexing; moved relevant methods from Data
